requirements.txt
numpy
PySide
pyopengl
# more?
pygments
